feat: Parsing chapters and storing them on Redis
The extraction process and the clean up work pretty well. There was a
timeout issue, but that has been rectified with the specification of a
longer `default_timeout`.
Another issue that is being faced is the failure upon visiting
particular links. Still needs to be investigated for the cause.

Error management is something that I will be looking into in a future
commit.