feat: Addition of the EPUB generator.
EPUB Generation is a good to go, I did not realize how small of a time
it takes to generate the epub, but this is very very fast. Albeit, there
might be hidden errors :)

We do face an issue in this with the zipfile module which tends to
create duplicate files.
This issue has also been raised in the following repository which also
utilizes the zipfile module.
- link: https://github.com/log2timeline/plaso/issues/8